# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Book Vault API
  version: 1.0.1
  description: |
    RESTful API for Book Vault audiobook library.

    ## Authentication

    **IMPORTANT**: This API supports **dual authentication** for maximum flexibility.
    All endpoints accept BOTH session cookies (web) AND Bearer tokens (mobile).

    ### Web Authentication (Session-based)
    - Login via NextAuth.js at `/api/auth/signin`
    - Session maintained via HTTP-only cookies
    - All documented endpoints support session auth

    ### Mobile Authentication (Bearer tokens)
    - Mobile-specific endpoints at `/api/auth/mobile/*`
    - Use JWT Bearer tokens: `Authorization: Bearer {token}`
    - All endpoints support both auth methods
    - See "Mobile Authentication" tag for mobile-specific endpoints

    ## Base URL
    - Development: http://localhost:3000
    - Production: TBD (AWS deployment)
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.bookvault.com
    description: Production server (TBD)
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (HTTP-only)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile JWT token (limited endpoint support)
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message
          example: 'Book not found'
    Pagination:
      type: object
      required: [page, limit, total, pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        pages:
          type: integer
          minimum: 0
          example: 8
          description: Total number of pages

    # Core models (from lib/types.ts)
    Book:
      type: object
      required: [id, asin, title, runtimeMinutes, coverUrl, audioUrl, authors]
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        asin:
          type: string
          example: 'B002V5D1CG'
        title:
          type: string
          example: 'The Name of the Wind'
        description:
          type: string
          nullable: true
          example: "Told in Kvothe's own voice, this is the tale of the magically gifted young man..."
        runtimeMinutes:
          type: integer
          minimum: 0
          example: 1068
        releaseDate:
          type: string
          format: date
          nullable: true
          example: '2008-04-01'
        publisher:
          type: string
          nullable: true
          example: 'DAW'
        coverUrl:
          type: string
          format: uri
          example: 'https://book-vault-media.s3.amazonaws.com/covers/B002V5D1CG.jpg'
        audioUrl:
          type: string
          format: uri
          example: 'https://book-vault-media.s3.amazonaws.com/audio/B002V5D1CG.m4b'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        narrators:
          type: array
          items:
            $ref: '#/components/schemas/Narrator'
        series:
          type: array
          items:
            $ref: '#/components/schemas/SeriesInfo'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    Author:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Patrick Rothfuss'
        asin:
          type: string
          nullable: true
          example: 'B002BMAYSQ'
    Narrator:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Nick Podehl'
        asin:
          type: string
          nullable: true
    SeriesInfo:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: 'The Kingkiller Chronicle'
        sequence:
          type: string
          nullable: true
          description: "Book number in series (can be '1', '1.5', 'Prequel', etc.)"
          example: '1'
        asin:
          type: string
          nullable: true
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Fantasy'
    Chapter:
      type: object
      required: [id, title, startTime, endTime, index]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: 'Chapter 1: A Place for Demons'
        startTime:
          type: number
          format: double
          minimum: 0
          example: 0
        endTime:
          type: number
          format: double
          minimum: 0
          example: 1806.5
        index:
          type: integer
          minimum: 0
          example: 0
...
    CategoryWithBookCount:
      type: object
      required: [id, name, bookCount]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Fantasy'
        level:
          type: integer
          minimum: 0
          example: 1
        bookCount:
          type: integer
          minimum: 0
          example: 150

# Security applied to all endpoints by default (session-based auth for web)
security:
  - sessionAuth: []
paths:
...
  /api/books:
    get:
      tags: [Books]
      summary: Get all books (paginated and sortable)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, author, narrator, series]
            default: title
          description: Sort order (title, author, narrator, or series)
      responses:
        '200':
          description: Paginated list of books
          content:
            application/json:
              schema:
                type: object
                required: [books, pagination]
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    type: object
                    required: [page, limit, total, pages]
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 150
                      pages:
                        type: integer
                        example: 8
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/books/{id}/chapters:
    get:
      tags: [Chapters]
      summary: Get chapters for a book (lazy extraction from audio file)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Book ID
      responses:
        '200':
          description: List of chapters
          content:
            application/json:
              schema:
                type: object
                required: [chapters, source]
                properties:
                  chapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chapter'
                  source:
                    type: string
                    enum: [database, extracted, none, unavailable, error]
                    description: |
                      Where chapters came from:
                      - database: Previously extracted chapters from DB
                      - extracted: Newly extracted from audio file
                      - none: No chapters found in audio file
                      - unavailable: ffprobe not installed (graceful degradation)
                      - error: Extraction failed
                    example: 'database'
                  message:
                    type: string
                    nullable: true
                    example: 'No chapters found in audio file'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Progress endpoints
...
  /api/browse/categories:
    get:
      tags: [Browse]
      summary: List all categories with book counts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of categories
          content:
            application/json:
              schema:
                type: object
                required: [results, pagination]
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryWithBookCount'
                  pagination:
                    type: object
                    required: [page, limit, total, pages]
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

  # Browse endpoints - Detail views (books by entity)
...
  /api/categories/{id}:
    get:
      tags: [Browse]
      summary: Get category details and books in the category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Category ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Category with books
          content:
            application/json:
              schema:
                type: object
                required: [id, name, books, pagination]
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                    example: 'Fantasy'
                  level:
                    type: integer
                    minimum: 0
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Library endpoints
...